requirements.txt
astroid==3.2.4
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2019.9.11
cffi==1.15.0
chardet==3.0.4
cryptography==35.0.0
defusedxml==0.6.0
diff-match-patch==20181111
dill==0.4.0
Django==2.2.28
django-allauth==0.46.0
django-autocomplete-light==3.4.1
django-crispy-forms==1.8.0
django-extensions==3.1.1
django-filter==2.2.0
django-formtools==2.1
django-haystack==2.8.1
django-import-export==1.2.0
django-menu-generator==1.0.4
django-summernote==0.8.11.6
django-tables2==2.1.1
django-taggit==1.2.0
django-webpack-loader==0.6.0
dotenv==0.9.9
et-xmlfile==1.0.1
gunicorn==19.9.0
idna==2.8
importlib-resources==5.4.0
isort==5.13.2
jdcal==1.4.1
MarkupPy==1.14
mccabe==0.7.0
oauthlib==3.1.1
odfpy==1.4.0
openpyxl==3.0.0
packaging==25.0
platformdirs==4.3.6
psycopg2-binary<2.9
pycparser==2.21
PyJWT==2.3.0
pylint==3.2.7
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
python-dotenv==1.0.1
python-memcached==1.59
python3-openid==3.2.0
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.3.0
sentry-sdk==0.13.1
setuptools-scm==6.4.2
six==1.12.0
soupsieve==2.7
sqlparse==0.3.0
tablib==0.14.0
tomli==2.2.1
tomlkit==0.13.3
tqdm==4.64.1
typing_extensions==4.13.2
urllib3==1.25.6
Whoosh==2.7.4
xlrd==1.2.0
xlwt==1.3.0
zipp==3.6.0
